Add JSONArray
