+++ django-stubs/db/models/functions/__init__.pyi
@@ -2,7 +2,6 @@ from .comparison import Cast as Cast
 from .comparison import Coalesce as Coalesce
 from .comparison import Collate as Collate
 from .comparison import Greatest as Greatest
-from .comparison import JSONObject as JSONObject
 from .comparison import Least as Least
 from .comparison import NullIf as NullIf
 from .datetime import Extract as Extract
@@ -29,6 +28,8 @@ from .datetime import TruncSecond as TruncSecond
 from .datetime import TruncTime as TruncTime
 from .datetime import TruncWeek as TruncWeek
 from .datetime import TruncYear as TruncYear
+from .json import JSONArray as JSONArray
+from .json import JSONObject as JSONObject
 from .math import Abs as Abs
 from .math import ACos as ACos
 from .math import ASin as ASin
